+++ src/Github/EnterpriseClient.php
@@ -8,17 +8,17 @@ namespace QL\Hal\Github;
 
 use Github\Client;
 use Github\HttpClient\Plugin\PathPrepend;
+use function GuzzleHttp\Psr7\uri_for;
 use Http\Client\Common\Plugin\AddHostPlugin;
-use Http\Discovery\UriFactoryDiscovery;
 
 class EnterpriseClient extends Client
 {
     public function __construct($httpClientBuilder, $apiVersion, $enterpriseUrl)
     {
-        parent::__construct($httpClientBuilder, $apiVersion, $enterpriseUrl);
+        parent::__construct($httpClientBuilder, $apiVersion);
         $httpClientBuilder->removePlugin(AddHostPlugin::class);
         $httpClientBuilder->removePlugin(PathPrepend::class);
-        $httpClientBuilder->addPlugin(new AddHostPlugin(UriFactoryDiscovery::find()->createUri($enterpriseUrl)));
+        $httpClientBuilder->addPlugin(new AddHostPlugin(uri_for($enterpriseUrl)));
         $httpClientBuilder->addPlugin(new PathPrepend(sprintf('/api/%s', $this->getApiVersion())));
     }
 }